% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asymptoticMultipleOutcomes.R
\name{asymptoticMultipleOutcomes}
\alias{asymptoticMultipleOutcomes}
\title{Compute asymptotic variance-covariance matrix of parameters in given models
It is used for models where bootstrap is not needed}
\usage{
asymptoticMultipleOutcomes(
  ...,
  data,
  family,
  data_index = NULL,
  score_epsilon = 1e-06
)
}
\description{
Compute asymptotic variance-covariance matrix of parameters in given models
It is used for models where bootstrap is not needed
}
