% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapMultipleOutcomes.R
\name{bootstrapMultipleOutcomes}
\alias{bootstrapMultipleOutcomes}
\title{Compute bootstrapped variance-covariance matrix of parameters in given models
It is used when at least one of the specified model needs bootstrap,
for example, Kaplan-Merier estimate for probability of survival, or
quantiles are used for prognostic variables.}
\usage{
bootstrapMultipleOutcomes(
  ...,
  data,
  data_index = NULL,
  nboot = 10,
  compute_cov = TRUE,
  seed = NULL
)
}
\description{
Compute bootstrapped variance-covariance matrix of parameters in given models
It is used when at least one of the specified model needs bootstrap,
for example, Kaplan-Merier estimate for probability of survival, or
quantiles are used for prognostic variables.
}
