% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkBootstrapInput.R
\name{checkBootstrapInput}
\alias{checkBootstrapInput}
\title{Process inputs of \code{multipleOutcomes} when bootstrap will be used to estimate
variance-covariance matrix}
\usage{
checkBootstrapInput(..., data, data_index)
}
\description{
Process inputs of \code{multipleOutcomes} when bootstrap will be used to estimate
variance-covariance matrix
}
