% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkDefaultInput.R
\name{checkDefaultInput}
\alias{checkDefaultInput}
\title{Process inputs of \code{multipleOutcomes} when asymptotic properties are used
to estimate variance-covariance matrix}
\usage{
checkDefaultInput(..., family, data, data_index)
}
\description{
Process inputs of \code{multipleOutcomes} when asymptotic properties are used
to estimate variance-covariance matrix
}
