% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparePointwiseConfidenceIntervalWidth.R
\name{comparePointwiseConfidenceIntervalWidth}
\alias{comparePointwiseConfidenceIntervalWidth}
\title{Generate two curves of survival probability with pointwise 95\% confidence
interval:
\enumerate{
\item PATED adjusted KM curve
\item Conventional KM curve
}}
\usage{
comparePointwiseConfidenceIntervalWidth(
  pated_res,
  km_res,
  transform = "identity"
)
}
\description{
Generate two curves of survival probability with pointwise 95\% confidence
interval:
\enumerate{
\item PATED adjusted KM curve
\item Conventional KM curve
}
}
