% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxph_.R
\name{coxph_}
\alias{coxph_}
\title{Creating Objects of Proportional Hazards Regression Model}
\usage{
coxph_(formula, data_index)
}
\arguments{
\item{formula}{see \code{formula} in \code{survival::coxph}.}

\item{data_index}{integer. Index of the data frame in the \code{data} argument of
\code{jointCovariance} to be used when fitting a proportional hazards model.}
}
\description{
\code{coxph_} is a wrapper function of \code{survival::coxph} to create an object to be
passed into \code{jointCovariance}, the main function of this package through its
argument \code{...}. The object defines how a proportional hazard model would be
fitted.
}
\details{
Not all arguments of \code{survival::coxph} are supported in \code{coxph_} due to the
complexity in handling environment and scope, which is particularly difficult
for arguments like \code{weights}, \code{subset}, etc.
}
