% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxphMO.R
\name{coxphMO}
\alias{coxphMO}
\title{return estimate of log HR from coxph model
to be used when bootstrap is needed.}
\usage{
coxphMO(formula, data = NULL, ties = c("efron", "breslow", "exact"))
}
\description{
return estimate of log HR from coxph model
to be used when bootstrap is needed.
}
