% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createKaplanMeierCurve.R
\name{createKaplanMeierCurve}
\alias{createKaplanMeierCurve}
\title{Create curve of survival probability based on conventional KM method, or
PATED adjusted KM estimates.
Refer to km_res or pated_res in \code{pated} about the format of \code{input}.
Transformations are supported when computing confidence intervals.
Refer to https://documentation.sas.com/doc/en/pgmsascdc/9.4_3.4/statug/statug_lifetest_details08.htm
or the LIFETEST proc in SAS for more details of transformation.}
\usage{
createKaplanMeierCurve(input, title = "", transform = "identity")
}
\description{
Create curve of survival probability based on conventional KM method, or
PATED adjusted KM estimates.
Refer to km_res or pated_res in \code{pated} about the format of \code{input}.
Transformations are supported when computing confidence intervals.
Refer to https://documentation.sas.com/doc/en/pgmsascdc/9.4_3.4/statug/statug_lifetest_details08.htm
or the LIFETEST proc in SAS for more details of transformation.
}
