% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractKaplanMeierTimes.R
\name{extractKaplanMeierTimes}
\alias{extractKaplanMeierTimes}
\title{Figure out the time points with at least one event.
This is used to generate KM-like curves.}
\usage{
extractKaplanMeierTimes(coef, sort = TRUE)
}
\description{
Figure out the time points with at least one event.
This is used to generate KM-like curves.
}
