% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gee_.R
\name{gee_}
\alias{gee_}
\title{Creating Objects of Generalized Estimation Equation Model}
\usage{
gee_(formula, id, family, corstr, R = NULL, b = NULL, Mv = 1, data_index = 1)
}
\arguments{
\item{formula}{see \code{formula} in \code{gee::gee}.}

\item{id}{character. Column name where clusters are defined. See \code{id} and
the Details section of \code{gee::gee}.}

\item{family}{see \code{family} in \code{gee::gee}.}

\item{corstr}{see \code{corstr} in \code{gee::gee}.}

\item{R}{see \code{R} in \code{gee::gee}.}

\item{b}{see \code{b} in \code{gee::gee}.}

\item{Mv}{see \code{Mv} in \code{gee::gee}.}

\item{data_index}{integer. Index of the data frame in the \code{data} argument of
\code{jointCovariance} to be used when fitting a GEE model.}
}
\description{
\code{gee_} is a wrapper function of \code{gee::gee} to create an object to be passed
into \code{jointCovariance}, the main function of this package through its
argument \code{...}. The object defines how a GEE model would be fitted.

This package does not import the package \code{gee}. Instead, codes of \code{gee} are
modified and integrated to compute score and information matrix. Thus, users
does not need to install the package \code{gee} to use this package.
}
\details{
Not all arguments of \code{stats::gee} are supported in \code{gee_} due to the
complexity in handling environment and scope, which is particularly difficult
for arguments like \code{subset}, etc.
}
