% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gFunction.R
\name{gFunction}
\alias{gFunction}
\title{g(S(t)), a transformed estimate of survival probability.
See https://documentation.sas.com/doc/en/pgmsascdc/9.4_3.4/statug/statug_lifetest_details08.htm}
\usage{
gFunction(conf.type = c("log", "log-log", "plain", "logit"))
}
\description{
g(S(t)), a transformed estimate of survival probability.
See https://documentation.sas.com/doc/en/pgmsascdc/9.4_3.4/statug/statug_lifetest_details08.htm
}
