% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gInverseFunction.R
\name{gInverseFunction}
\alias{gInverseFunction}
\title{Inversed function of g(S(t)).
See https://documentation.sas.com/doc/en/pgmsascdc/9.4_3.4/statug/statug_lifetest_details08.htm}
\usage{
gInverseFunction(conf.type = c("log", "log-log", "plain", "logit"))
}
\description{
Inversed function of g(S(t)).
See https://documentation.sas.com/doc/en/pgmsascdc/9.4_3.4/statug/statug_lifetest_details08.htm
}
