% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_.R
\name{glm_}
\alias{glm_}
\title{Creating Objects of Generalized Linear Models}
\usage{
glm_(formula, family, data_index)
}
\arguments{
\item{formula}{see \code{formula} in \code{stats::glm}.}

\item{family}{currently supports \code{"gaussian"} or \code{"binomial"}. Other families
are under testing.}

\item{data_index}{integer. Index of the data frame in the \code{data} argument of
\code{jointCovariance} to be used when fitting a generalized linear model.}
}
\description{
\code{glm_} is a wrapper function of \code{stats::glm} to create an object to be passed
into \code{jointCovariance}, the main function of this package through its
argument \code{...}. The object defines how a GLM model would be fitted.
}
\details{
Not all arguments of \code{stats::glm} are supported in \code{glm_} due to the
complexity in handling environment and scope, which is particularly difficult
for arguments like \code{weights}, \code{subset}, etc.
}
