% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IDMapping.R
\name{IDMapping}
\alias{IDMapping}
\title{df is of length that equals to the number of specified models
df\link{i} is the number of parameters in model i
This function returns index of parameters of model i in the vector of
parameters of all models}
\usage{
IDMapping(df, i, var_name = NULL)
}
\description{
df is of length that equals to the number of specified models
df\link{i} is the number of parameters in model i
This function returns index of parameters of model i in the vector of
parameters of all models
}
