% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inverseHessianMatrixGlm.R
\name{inverseHessianMatrixGlm}
\alias{inverseHessianMatrixGlm}
\title{glm_fit is object returned from glm()}
\usage{
inverseHessianMatrixGlm(glm_fit)
}
\description{
glm_fit is object returned from glm()
}
