% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/km_.R
\name{km_}
\alias{km_}
\title{Creating Objects of Kaplan-Meier Curve}
\usage{
km_(formula, times = NULL, conf_type, data_index = 1)
}
\arguments{
\item{formula}{a formula created by \code{survival::Surv()}.}

\item{times}{numeric vector of time. Survival probabilities at \code{times} are
computed (with g-transformation defined by \code{conf_type}).}

\item{conf_type}{character. Type of confidence interval. It must be one of
\code{"log"}, \code{"log-log"}, \code{"plain"}, \code{"logit"}, or \code{"arcsin"}.}

\item{data_index}{integer. Index of the data frame in the \code{data} argument of
\code{jointCovariance} to be used when fitting a generalized linear model.}
}
\description{
\code{km_} is a wrapper function creating an object of Kaplan-Meier curve to be
passed into \code{jointCovariance}, the main function of this package through its
argument \code{...}. The object defines how a Kaplan-Meier curve would be fitted.
}
\details{
Usually, g-transformation is applied to the survival probability \code{S(t)}
to obtain pointwise confidence interval of a Kaplan-Meier curve. This
can be achieved by specifying \code{conf_type}. For identity transformation,
use \code{conf_type = "plain"}.

This function can only be used with \code{jointCovariance} when the bootstrap
method is used to estimate variance-covariance matrix of multiple outcome
models.
}
