% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmMO.R
\name{kmMO}
\alias{kmMO}
\title{xxx}
\usage{
kmMO(formula, conf.type, times = NULL, data = NULL)
}
\arguments{
\item{formula}{a survival object created by \code{survival::Surv()}.}

\item{conf.type}{type of confidence interval. It must be one of \code{"log"},
\code{"log-log"}, \code{"plain"}, \code{"logit"}, or \code{"arcsin"}.}

\item{times}{numeric. A vector of times to compute the survival probabilities.
By default, i.e. \code{NULL}, jump times in data is used.}

\item{data}{a data frame. User should use its default value \code{NULL}.}
}
\description{
xxx
}
