% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logrank_.R
\name{logrank_}
\alias{logrank_}
\title{Creating Objects of Logrank Test}
\usage{
logrank_(formula, ties = c("efron", "breslow", "exact"), data_index = 1)
}
\arguments{
\item{formula}{see \code{formula} in \code{survival::coxph}.}

\item{data_index}{integer. Index of the data frame in the \code{data} argument of
\code{jointCovariance} to be used when computing testing statistic of logrank
test.}
}
\description{
\code{logrank_} is a wrapper function \code{survival::coxph} to create an object to be
passed into \code{jointCovariance}, the main function of this package through its
argument \code{...}. Logrank test is the score test under the proportional hazards
regression model. The object defines how a logrank test would be computed.
}
\details{
Not all arguments of \code{survival::coxph} are supported in \code{logrank_} due to the
complexity in handling environment and scope, which is particularly difficult
for arguments like \code{weights}, \code{subset}, etc.
}
