% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logrankMO.R
\name{logrankMO}
\alias{logrankMO}
\title{xxx}
\usage{
logrankMO(formula, data = NULL, ties = c("efron", "breslow", "exact"))
}
\arguments{
\item{formula}{a survival object using function \code{survival::Surv()}.}

\item{data}{a data frame. User should always use its default value \code{NULL}.}

\item{ties}{method to handle ties, must be \code{"efron"}, \code{"breslow"},
or \code{"exact"}.}
}
\description{
xxx
}
