% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmrm_.R
\name{mmrm_}
\alias{mmrm_}
\title{Creating Objects of Mixed Models for Repeated Measures}
\usage{
mmrm_(
  formula,
  covariance = NULL,
  reml = TRUE,
  control = mmrm::mmrm_control(...),
  ...,
  data_index = 1
)
}
\arguments{
\item{formula}{see \code{formula} in \code{mmrm::mmrm}.}

\item{covariance}{see \code{covariance} in \code{mmrm::mmrm}.}

\item{reml}{see \code{reml} in \code{mmrm::mmrm}.}

\item{control}{see \code{control} in \code{mmrm::mmrm}.}

\item{...}{see \code{...} in \code{mmrm::mmrm}.}

\item{data_index}{integer. Index of the data frame in the \code{data} argument of
\code{jointCovariance} to be used when fitting a GEE model.}
}
\description{
\code{mmrm_} is a wrapper function of \code{mmrm::mmrm} to create an object to be passed
into \code{jointCovariance}, the main function of this package through its
argument \code{...}. The object defines how a MMRM model would be fitted.
}
\details{
The argument \code{weights} of \code{mmrm::mmrm} is supported in \code{mmrm_} due to the
complexity in handling environment and scope.

Please always refer to help document of \code{mmrm::mmrm} before using \code{mmrm_}.
For example, time variable and observation ID must be factor variables in
some cases, otherwise error may be prompted. Users can call \code{mmrm::mmrm}
using the same arguments being passed to \code{mmrm_} to check validity.
}
