% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pated.R
\name{plot.pated}
\alias{plot.pated}
\title{Plot PATED Analysis Results}
\usage{
\method{plot}{pated}(x, ...)
}
\arguments{
\item{x}{an object returned from \code{pated()}.}

\item{...}{currently not supported.}
}
\value{
\code{NULL}
}
\description{
Plot PATED Analysis Results
}
