% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summary.jointCovariance.R
\name{print.summary.jointCovariance}
\alias{print.summary.jointCovariance}
\title{Title Summarize an Analysis of Multiple Outcomes.}
\usage{
\method{print}{summary.jointCovariance}(x, ...)
}
\arguments{
\item{x}{an object returned by \code{jointCovariance()}.}

\item{...}{for debugging only.}
}
\value{
an invisible object.
}
\description{
Summarize an analysis of multiple outcomes.
}
\examples{
## no example
}
