% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summary.multipleOutcomes.R
\name{print.summary.multipleOutcomes}
\alias{print.summary.multipleOutcomes}
\title{Title}
\usage{
\method{print}{summary.multipleOutcomes}(x, ...)
}
\arguments{
\item{x}{an object returned by \code{multipleOutcomes()}.}

\item{...}{for debugging only.}
}
\value{
an invisible object.
}
\description{
Title
}
\examples{
## no example
}
