% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantileMO.R
\name{quantileMO}
\alias{quantileMO}
\title{Return difference in quantile between arms
formula can be endpoint ~ trt}
\usage{
quantileMO(formula, data = NULL, probs = c(0.25, 0.5, 0.75))
}
\description{
Return difference in quantile between arms
formula can be endpoint ~ trt
}
