% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleWithReplacement.R
\name{sampleWithReplacement}
\alias{sampleWithReplacement}
\title{Generate bootstrap dataset.
When missing data presents, dataset is split into groups. Patients in
the same group have missing on the same covariates. Bootstrap is carried
out in each of the groups}
\usage{
sampleWithReplacement(data)
}
\description{
Generate bootstrap dataset.
When missing data presents, dataset is split into groups. Patients in
the same group have missing on the same covariates. Bootstrap is carried
out in each of the groups
}
