% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateMoData.R
\name{simulateMoData}
\alias{simulateMoData}
\title{Generating Data for Simulation and Testing}
\usage{
simulateMoData(n = 500, hr = 0.8, seed = NULL)
}
\arguments{
\item{n}{an integer for total sample size of a randomized control trial of
two arms.}

\item{hr}{hazard ratio of treatment.}

\item{seed}{random seed. By default \code{NULL} for no seed being specified.}
}
\description{
\code{simulateMoData} generates data for simulation and testing purposes.
}
