% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.jointCovariance.R
\name{summary.jointCovariance}
\alias{summary.jointCovariance}
\title{Object Summaries}
\usage{
\method{summary}{jointCovariance}(object, model_index = NULL, ...)
}
\arguments{
\item{object}{an object returned by \code{jointCovariance()}.}

\item{model_index}{\code{NULL} if displaying summary of all fitted models; otherwise,
an integer indicating the fitted model.}

\item{...}{for debugging only}
}
\value{
a list
}
\description{
\code{summary} method for class \code{jointCovariance}.
}
