% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.multipleOutcomes.R
\name{summary.multipleOutcomes}
\alias{summary.multipleOutcomes}
\title{Object Summaries}
\usage{
\method{summary}{multipleOutcomes}(object, model_index = NULL, ...)
}
\arguments{
\item{object}{an object returned by \code{multipleOutcomes()}.}

\item{model_index}{\code{NULL} if displaying summary of all fitted models; otherwise,
an integer indicating the fitted model.}

\item{...}{for debugging only}
}
\value{
a list
}
\description{
\code{summary} method for class \code{multipleOutcomes}.
}
