% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcov.jointCovariance.R
\name{vcov.jointCovariance}
\alias{vcov.jointCovariance}
\title{Calculate Variance-Covariance Matrix for a Fitted Model Object}
\usage{
\method{vcov}{jointCovariance}(object, model_index = NULL, ...)
}
\arguments{
\item{object}{an object returned by \code{jointCovariance()}.}

\item{model_index}{\code{NULL} if displaying covariance matrix of all fitted models;
otherwise, an integer indicating the fitted model.}

\item{...}{for debugging only}
}
\value{
a matrix of covariance of all estimates
}
\description{
Returns the variance-covariance matrix of the main parameters of fitted model
objects. The "main" parameters of models correspond to those returned by \code{coef}.
}
