% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcov.multipleOutcomes.R
\name{vcov.multipleOutcomes}
\alias{vcov.multipleOutcomes}
\title{Calculate Variance-Covariance Matrix for a Fitted Model Object}
\usage{
\method{vcov}{multipleOutcomes}(object, model_index = NULL, ...)
}
\arguments{
\item{object}{an object returned by \code{multipleOutcomes()}.}

\item{model_index}{\code{NULL} if displaying covariance matrix of all fitted models;
otherwise, an integer indicating the fitted model.}

\item{...}{for debugging only}
}
\value{
a matrix of covariance of all estimates
}
\description{
Calculate Variance-Covariance Matrix for a Fitted Model Object
}
