% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geeMO.R
\name{geeMO}
\alias{geeMO}
\title{return estimates from GEE model
to be used when bootstrap is needed.}
\usage{
geeMO(formula, id, data = NULL, family, corstr)
}
\description{
return estimates from GEE model
to be used when bootstrap is needed.
}
