% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmMO.R
\name{glmMO}
\alias{glmMO}
\title{return estimates from GLM model
to be used when bootstrap is needed.}
\usage{
glmMO(formula, family, data = NULL)
}
\description{
return estimates from GLM model
to be used when bootstrap is needed.
}
