% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pated.R
\name{pated}
\alias{pated}
\title{Prognostic Variables Assisted Treatment Effect Detection}
\usage{
pated(
  ...,
  data,
  nboot = 0,
  compute_cov = TRUE,
  seed = NULL,
  transform = "identity"
)
}
\arguments{
\item{...}{formulas of models to be fitted, or moment functions for gmm.}

\item{data}{a data frame if all models are fitted on the same dataset;
otherwise a list of data frames for fitting models in \code{...}. Note that a
dataset can be used to fit multiple models, thus, \code{length(data)} is unnecessary
to be equal to the number of models in \code{...}. The row names in a data frame
are treated as sample IDs. Consequently, for any two records in different
data frames that correspond to the same sample, their row names should be
consistent.}

\item{nboot}{non-zero integer if bootstrap is adopted. By default 0.}

\item{compute_cov}{logic. If \code{TRUE}, empirical covariance matrix is computed
using bootstrap estimate and returned. Bootstrap estimate will be abandoned. If
\code{FALSE}, bootstrap estimate will be returned and no empirical covariance
matrix is computed.}

\item{seed}{random seed when generate bootstrap data.}

\item{transform}{character. Now only supports \code{"identity"}.}

\item{family}{a character vector of families to be used in the models.
All families supported by \code{multipleOutcomes} are also supported by \code{pated}.
\code{family} can be of length 1 if all models are fitted in the same family;
otherwise family should be specified for each of the models in \code{...}.}

\item{data_index}{\code{NULL} if \code{data} is a data frame; otherwise, a vector in
integer specifying mapping a model in \code{...} to a data frame in \code{data} (a list).}
}
\value{
a data frame of testing results.
}
\description{
\code{pated} is a wrapper function of \code{multipleOutcomes} for testing treatment effect
in randomized clinical trials. It assumes that prognostic variables are fully
randomized. This assumption can help enhancing statistical power of conventional
approaches in detecting the treatment effect. Specifically, the sensitivity
of the conventional models specified in \code{...} are improved by \code{pated}.
}
\examples{
## More examples can be found in the vignettes.
library(survival)
library(mvtnorm)

genData <- function(seed = NULL){
  set.seed(seed)
  n <- 300
  sigma <- matrix(c(1, .6, .6, 1), 2)
  x <- rmvnorm(n, sigma = sigma)
  z1 <- rbinom(n, 1, .6)
  z2 <- rnorm(n)
  trt <- rbinom(n, 1, .5)

  bet <- c(-.2, .2)
  y <- -.5 + x \%*\% bet + z1 * .3 - z2 * .1 + .1 * trt - .1 * rnorm(n)
  death <- rbinom(n, 1, .8)
  data.frame(
    y = as.numeric(y), trt = trt,
    z1 = z1, z2 = z2,
    x1 = x[, 1], x2 = x[, 2],
    death, pid = paste0('s-', seq_len(n))
  )
}

dat <- genData(seed = 31415926)

fit <-
  pated(
    coxph_(Surv(time = y, event = death) ~ trt, data_index = 1),
    glm_(z1 ~ trt, family = 'binomial',  data_index = 1),
    glm_(z2 ~ trt, family = 'gaussian',  data_index = 1),
    glm_(x1 ~ trt, family = 'gaussian',  data_index = 1),
    glm_(x2 ~ trt, family = 'gaussian',  data_index = 1),
    data = list(dat)
  )

fit

}
