% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pated.R
\name{pated}
\alias{pated}
\title{Prognostic Variables Assisted Treatment Effect Detection}
\usage{
pated(..., family, data)
}
\arguments{
\item{...}{x}

\item{family}{x}

\item{data}{x}
}
\value{
x
}
\description{
Prognostic Variables Assisted Treatment Effect Detection
}
\examples{
# x
}
