% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pated.R
\name{pated}
\alias{pated}
\title{Prognostic Variables Assisted Treatment Effect Detection}
\usage{
pated(..., family, data, data_index = NULL)
}
\arguments{
\item{...}{formulas of models to be fitted, or moment functions for gmm.}

\item{family}{a character vector of families to be used in the models.
All families supported by \code{multipleOutcomes} are also supported by \code{pated}.
\code{family} can be of length 1 if all models are fitted in the same family;
otherwise family should be specified for each of the models in \code{...}.}

\item{data}{a data frame if all models are fitted on the same dataset;
otherwise a list of data frames for fitting models in \code{...}. Note that a
dataset can be used to fit multiple models, thus, \code{length(data)} is unnecessary
to be equal to the number of models in \code{...}. The row names in a data frame
are treated as sample IDs. Consequently, for any two records in different
data frames that correspond to the same sample, their row names should be
consistent.}

\item{data_index}{\code{NULL} if \code{data} is a data frame; otherwise, a vector in
integer specifying mapping a model in \code{...} to a data frame in \code{data} (a list).}
}
\value{
a data frame of testing results.
}
\description{
\code{pated} is a wrapper function of \code{multipleOutcomes} for testing treatment effect
in randomized clinical trials. It assumes that prognostic variables are fully
randomized. This assumption can help enhancing statistical power of conventional
approaches in detecting the treatment effect. Specifically, the sensitivity
of the conventional models specified in \code{...} are improved by \code{pated}.
}
\examples{
## More examples can be found in the vignettes.
library(survival)
library(mvtnorm)
library(tidyr)
genData <- function(seed = NULL){

  set.seed(seed)
  n <- 200
  sigma <- matrix(c(1, .6, .6, 1), 2)
  x <- rmvnorm(n, sigma = sigma)
  z1 <- rbinom(n, 1, .6)
  z2 <- rnorm(n)
  gam <- c(.1, -.2)
  trt <- rbinom(n, 1, .5)

  bet <- c(-.2,.2)
  y <- -.5+x \%*\% bet + z1 * .3 - z2 * .1 + .1 * trt-.1 * rnorm(n)
  death <- rbinom(n, 1, .8)
  id <- 1:n
  data.frame(
    y = y, trt = trt, 
    z1 = z1, z2 = z2, 
    x1 = x[, 1], x2 = x[, 2], 
    death, id)

}

dat1 <- genData(seed = 31415926)

## create a dataset with repeated measurements x
dat2 <- dat1 \%>\% pivot_longer(c(x1, x2), names_to='tmp', values_to='x') \%>\% 
dplyr::select(x, trt, id) \%>\% as.data.frame()

fit <- 
  pated(
    Surv(time=y, event=death) ~ trt,
    z1 ~ trt, 
    z2 ~ trt, 
    x ~ trt, 
    family=c('logrank', 'binomial', 'gaussian', 'gee+id+gaussian'), 
    data=list(dat1, dat2), data_index = c(1, 1, 1, 2))

fit

}
